+++ python/pyspark/sql/tests/connect/test_connect_session.py
@@ -18,6 +18,7 @@
 import os
 import unittest
 import uuid
+from typing import Optional
 
 from pyspark.util import is_remote_only
 from pyspark.errors import PySparkException
@@ -32,13 +33,18 @@ from pyspark.testing.utils import timeout
 if should_test_connect:
     import grpc
     from pyspark.sql.connect.session import SparkSession as RemoteSparkSession
-    from pyspark.sql.connect.client import ChannelBuilder
+    from pyspark.sql.connect.client import ChannelBuilder, DefaultChannelBuilder
     from pyspark.errors.exceptions.connect import (
         AnalysisException,
         SparkConnectException,
         SparkConnectGrpcException,
         SparkUpgradeException,
     )
+
+    class CustomChannelBuilder(DefaultChannelBuilder):
+        @property
+        def userId(self) -> Optional[str]:
+            return "abc"
 
 
 @unittest.skipIf(is_remote_only(), "Session creation different from local mode")
@@ -297,6 +303,89 @@ class SparkConnectSessionTests(ReusedConnectTestCase):
             session.range(3).collect()
 
         self.assertTrue("Invalid authentication token" in str(e.exception))
+
+
+@unittest.skipIf(not should_test_connect, connect_requirement_message)
+class SparkConnectSessionBuilderTests(unittest.TestCase):
+    def setUp(self):
+        # Reset class-level session state so tests are order-independent.
+        RemoteSparkSession._default_session = None
+        RemoteSparkSession._active_session.session = None
+
+    def test_fails_to_create_session_without_remote_and_channel_builder(self):
+        with self.assertRaises(ValueError):
+            RemoteSparkSession.builder.getOrCreate()
+
+    def test_fails_to_create_when_both_remote_and_channel_builder_are_specified(self):
+        with self.assertRaises(ValueError):
+            (
+                RemoteSparkSession.builder.channelBuilder(CustomChannelBuilder("sc://localhost"))
+                .remote("sc://localhost")
+                .getOrCreate()
+            )
+
+    def test_creates_session_with_channel_builder(self):
+        test_session = RemoteSparkSession.builder.channelBuilder(
+            CustomChannelBuilder("sc://other")
+        ).getOrCreate()
+        host = test_session.client.host
+        # Skip release_session() since "sc://other" is a fake remote.
+        test_session.release_session_on_close = False
+        test_session.stop()
+
+        self.assertEqual("other", host)
+
+    def test_creates_session_with_remote(self):
+        test_session = RemoteSparkSession.builder.remote("sc://other").getOrCreate()
+        host = test_session.client.host
+        test_session.release_session_on_close = False
+        test_session.stop()
+
+        self.assertEqual("other", host)
+
+    def test_session_stop(self):
+        session = RemoteSparkSession.builder.remote("sc://other").getOrCreate()
+        session.release_session_on_close = False
+
+        self.assertFalse(session.is_stopped)
+        session.stop()
+        self.assertTrue(session.is_stopped)
+
+    def test_session_create_sets_active_session(self):
+        session = RemoteSparkSession.builder.remote("sc://abc").create()
+        session2 = RemoteSparkSession.builder.remote("sc://other").getOrCreate()
+
+        self.assertIs(session, session2)
+        session.release_session_on_close = False
+        session.stop()
+
+    def test_active_session_expires_when_client_closes(self):
+        s1 = RemoteSparkSession.builder.remote("sc://other").getOrCreate()
+        s2 = RemoteSparkSession.getActiveSession()
+
+        self.assertIs(s1, s2)
+
+        # We don't call close() to avoid executing ExecutePlanResponseReattachableIterator
+        s1._client._closed = True
+
+        self.assertIsNone(RemoteSparkSession.getActiveSession())
+        s3 = RemoteSparkSession.builder.remote("sc://other").getOrCreate()
+
+        self.assertIsNot(s1, s3)
+
+    def test_default_session_expires_when_client_closes(self):
+        s1 = RemoteSparkSession.builder.remote("sc://other").getOrCreate()
+        s2 = RemoteSparkSession._get_default_session()
+
+        self.assertIs(s1, s2)
+
+        # We don't call close() to avoid executing ExecutePlanResponseReattachableIterator
+        s1._client._closed = True
+
+        self.assertIsNone(RemoteSparkSession._get_default_session())
+        s3 = RemoteSparkSession.builder.remote("sc://other").getOrCreate()
+
+        self.assertIsNot(s1, s3)
 
 
 @unittest.skipIf(not should_test_connect, connect_requirement_message)
